Move to golangci-lint
This is faster and does more checks.

* Remove home-grown `lint` script.
* Install the `golangci-lint` binary in the build container.
* Remove individual tools `gocyclo`, `golint`, `errcheck` which are included in `golangci-lint`.

Run in `--new-from-rev` mode so it doesn't fail the build on a bunch
of warnings that were pre-existing.

Still running `go vet` separately because I couldn't see that the `vet`
tool inside `golangci-lint` is the same thing.

Also run `misspell` specifically on the docs because `golangci-lint`
only does `*.go`.

Signed-off-by: Bryan Boreham <bryan@weave.works>